ai_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.20
langchain-community==0.0.38
chromadb==0.4.24
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
numpy==1.26.2

